rename endpoint files
