+++ lib/udap_security_test_kit/endpoints/mock_udap_server/registration_endpoint.rb
